Move enum into types
