+++ ui/src/shared/constants/index.tsx
@@ -487,9 +487,3 @@ export const HANDLE_PIXELS = 20
 export const MIN_HANDLE_PIXELS = 20
 export const MAX_SIZE = 1
 export const MIN_SIZE = 0
-
-// Write Data Modes
-export enum WriteDataMode {
-  Manual = 'Manual Entry',
-  File = 'File Upload',
-}
