+++ ui/src/types/dataExplorer.ts
@@ -1,0 +1,5 @@
+// Write Data Modes
+export enum WriteDataMode {
+  Manual = 'Manual Entry',
+  File = 'File Upload',
+}
